fix(LearningDiary): fixes not creation of LTB Pages when being in a lesson for at least one minute
