+++ apps/site/pages/courses/[courseSlug]/index.tsx
@@ -2,11 +2,9 @@ import { GetServerSideProps } from "next";
 import { MDXRemote } from "next-mdx-remote";
 import Image from "next/image";
 import Link from "next/link";
-import { useCallback, useMemo } from "react";
-
+import {  useMemo } from "react";
 import { PlayIcon, PlusCircleIcon } from "@heroicons/react/24/solid";
 import { LessonType } from "@prisma/client";
-import { trpc } from "@self-learning/api-client";
 import { useCourseCompletion } from "@self-learning/completion";
 import { database } from "@self-learning/database";
 import { useEnrollmentMutations, useEnrollments } from "@self-learning/enrollment";
@@ -18,7 +16,7 @@ import {
 	LessonInfo,
 	ResolvedValue
 } from "@self-learning/types";
-import { AuthorsList, useTimeout } from "@self-learning/ui/common";
+import { AuthorsList } from "@self-learning/ui/common";
 import * as ToC from "@self-learning/ui/course";
 import { CenteredContainer, CenteredSection } from "@self-learning/ui/layouts";
 import { formatDateAgo, formatSeconds } from "@self-learning/util/common";
@@ -162,27 +160,9 @@ async function getCourse(courseSlug: string) {
 	});
 }
 
-function useLearningDiaryRecording(course: CourseProps["course"]) {
-	const { mutateAsync: createLearningDiaryEntry } = trpc.learningDiary.create.useMutation();
-	const log = useCallback(async () => {
-		try {
-			await createLearningDiaryEntry({
-				courseSlug: course.slug
-			});
-		} catch (e) {}
-	}, [createLearningDiaryEntry, course.slug]);
-	useTimeout({ callback: log, delayInMilliseconds: 60000 });
-}
-
-function Writer({ course }: { course: CourseProps["course"] }) {
-	useLearningDiaryRecording(course);
-	return <></>;
-}
-
 export default function Course({ course, summary, content, markdownDescription }: CourseProps) {
 	return (
 		<div className="bg-gray-50 pb-32">
-			<Writer course={course} />
 			<CenteredSection className="bg-gray-50">
 				<CourseHeader course={course} content={content} summary={summary} />
 			</CenteredSection>
