+++ libs/ui/lesson/src/lib/playlist/playlist.tsx
@@ -6,11 +6,11 @@ import {
 	PlayIcon
 } from "@heroicons/react/24/solid";
 import { CourseCompletion, extractLessonIds, LessonMeta } from "@self-learning/types";
-import { Divider, ProgressBar } from "@self-learning/ui/common";
-import { motion } from "framer-motion";
+import { Divider, ProgressBar, useTimeout } from "@self-learning/ui/common";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { trpc } from "@self-learning/api-client";
 
 export type PlaylistChapter = {
 	title: string;
@@ -43,9 +43,25 @@ type PlaylistProps = {
 	completion?: CourseCompletion;
 };
 
+function useLearningDiaryRecording(courseSlug: string) {
+	const { mutateAsync: createLearningDiaryEntry } = trpc.learningDiary.create.useMutation();
+	const log = useCallback(async () => {
+		try {
+			await createLearningDiaryEntry({
+				courseSlug: courseSlug
+			});
+		} catch (e) {}
+	}, [createLearningDiaryEntry, courseSlug]);
+	useTimeout({ callback: log, delayInMilliseconds: 60000 });
+}
+
+function Writer({ courseSlug }: { courseSlug: string }) {
+	useLearningDiaryRecording(courseSlug);
+	return <></>;
+}
+
 export function Playlist({ content, course, lesson, completion }: PlaylistProps) {
 	const [contentWithCompletion, setContentWithCompletion] = useState(content);
-
 	useEffect(() => {
 		if (!completion) {
 			return;
@@ -62,6 +78,7 @@ export function Playlist({ content, course, lesson, completion }: PlaylistProps)
 
 	return (
 		<>
+			<Writer courseSlug={course.slug} />
 			<PlaylistHeader
 				content={content}
 				course={course}
